change: Change double types to decimal

+++ src/TaskrowSharp/Models/ClientModels/Client.cs
@@ -26,7 +26,7 @@ public class Client
     //public List<object> ClientSegment { get; set; } = [];
     public List<ClientAddress> ClientAddress { get; set; } = [];
     public ClientAdministrativeDetail ClientAdministrativeDetail { get; set; }
-    public double? BonusOrderPercentage { get; set; }
+    public decimal? BonusOrderPercentage { get; set; }
     //public object? HasOpenedJobs { get; set; }
     public bool IsSupplier { get; set; }
     public bool NotToSearchAsClient { get; set; }
